Hector

+++ public/javascripts/lists.js
@@ -48,41 +48,44 @@ async function fetchLists() {
         //create an anchor, rename btn, and delete btn to each list item
         lists.forEach(list => {
 
-            const anchor = document.createElement('anchor');
-            const btnRename = document.createElement('button');
-            const btnDelete = document.createElement('button');
-            const divContainer = document.createElement('div');
+            if (list.name !== "_hidden") {
 
-            divContainer.setAttribute('id', `div-container-${list.id}`);
+                const anchor = document.createElement('anchor');
+                const btnRename = document.createElement('button');
+                const btnDelete = document.createElement('button');
+                const divContainer = document.createElement('div');
 
-            anchor.setAttribute('id', list.id);
-            anchor.setAttribute('class', "list-anchors")
-            anchor.setAttribute('href', `http://localhost:8080/lists/${list.id}`);
-            btnRename.setAttribute('class', `list__btn--rename ${list.id}`);
-            btnDelete.setAttribute('class', `list__btn--delete ${list.id}`);
+                divContainer.setAttribute('id', `div-container-${list.id}`);
 
-            anchor.innerText = list.name;
-            btnDelete.innerText = "Delete";
+                anchor.setAttribute('id', list.id);
+                anchor.setAttribute('class', "list-anchors")
+                anchor.setAttribute('href', `http://localhost:8080/lists/${list.id}`);
+                btnRename.setAttribute('class', `list__btn--rename ${list.id}`);
+                btnDelete.setAttribute('class', `list__btn--delete ${list.id}`);
 
-            //Create the modal for the rename btn
-            btnRename.innerHTML = `Rename <div id="myModal-rename-${list.id}" class="modal-rename ${list.id}">
+                anchor.innerText = list.name;
+                btnDelete.innerText = "Delete";
+
+                //Create the modal for the rename btn
+                btnRename.innerHTML = `Rename <div id="myModal-rename-${list.id}" class="modal-rename ${list.id}">
                                         <div class="modal-content-rename">
                                             <form class="list-rename-form">
-                                                <span class="close-rename">X</span>
+                                                <span id="close-rename-${list.id}" class="close-rename">X</span>
                                                 <p class="add-list-rename">Rename List</p>
                                                 <label class="list-label-rename"> List name </label>
                                                 <br>
-                                                <input type="text" name="name" class="list-name-rename--input" required>
+                                                <input type="text" name="name" id="rename-input-${list.id}" class="list-name-rename--input" required>
                                                 <br>
                                                 <input type="submit" value="Save" class="list-rename">
                                                 <input type="button" value="Cancel" id="cancel-${list.id}" class="list-cancel-rename">
                                         </div>
                                     </div>`;
 
-            divLists.appendChild(divContainer);
-            divContainer.appendChild(anchor);
-            divContainer.appendChild(btnRename);
-            divContainer.appendChild(btnDelete);
+                divLists.appendChild(divContainer);
+                divContainer.appendChild(anchor);
+                divContainer.appendChild(btnRename);
+                divContainer.appendChild(btnDelete);
+            }
         });
 
         const inputCsurf = document.querySelector('#csurf')
@@ -93,114 +96,21 @@ async function fetchLists() {
 }
 
 //-------------------------------------------------------------------------------
-
-const tasksContainer = document.querySelector('#tasks-section__tasks-container');
-let selectedTaskIds = new Set();
 //Display all tasks from the targeted list
-//Display all summary of the targeted list
-// async function fetchOneList(id) {
+async function fetchOneList(id) {
 
-//     // const ulTasks = document.querySelector('#tasks');
-//     try {
-//         // const res = await fetch(`/lists/${id}`);
-//         // if (!res.ok) {
-//         //     throw res;
-//         // }
-//         // const { list, tasks } = await res.json();
-//         const res = await fetch(`/tasks?listIds=${[id]}`);
-//         if (!res.ok) {
-//             throw res;
-//         }
-//         const json = await res.json();
-//         console.log(json);
-//         // tasks.forEach(task => {
-//         //     const li = document.createElement('li');
-//         //     li.innerHTML = task.name;
-//         //     ulTasks.appendChild(li);
-//         // });
-//     }
-//     catch (err) {
-//         console.log(err);
-//     }
-// }
-
-
-const getTasks = async (listIds = []) => {
-    const res = await fetch(`/tasks?listIds=${[...listIds]}`, {
-        method: 'GET',
-        headers: {
-            "Content-Type": "application/json"
+    // const ulTasks = document.querySelector('#tasks');
+    try {
+        const res = await fetch(`/lists/${id}`);
+        if (!res.ok) {
+            throw res;
         }
-    });
-    if (!res.ok)
-        throw res;
-    const { tasks } = await res.json();
-    return tasks;
+        const { list, tasks } = await res.json();
+    }
+    catch (err) {
+        console.log(err);
+    }
 }
-
-const filterTasks = async (tasks, query) => {
-    return tasks.filter(task => {
-        for (let prop in query) {
-            // Include must store an object with props: term and from
-            if (prop === 'include') {
-                const { term, from } = query[prop];
-                if (!task[from].toLowerCase().includes(term.toLowerCase()))
-                    return false;
-            }
-            // Exclude must store an object with props: term and from
-            else if (prop === 'exclude') {
-                const { term, from } = query[prop];
-                if (task[from].toLowerCase().includes(term.toLowerCase()))
-                    return false;
-            }
-            // Check if a query prop/value matches a task prop/value
-            else if (task[prop] !== query[prop])
-                return false;
-        }
-        return true;
-    });
-}
-
-const displayTasks = async (tasks) => {
-    if (!tasks) tasks = await getTasks();
-
-    // Reset selectedTaskIds
-    selectedTaskIds = new Set();
-
-    const tasksHtml = tasks.map(task => {
-        let taskStr = '';
-
-        if (task.sets) taskStr += `${task.sets} `;
-        if (task.reps) taskStr += `x ${task.reps} `;
-        taskStr += task.name;
-        if (task.duration) {
-            const dur = convertSeconds(task.duration);
-            taskStr += ` for ${dur[0]}:${dur[1]}:${dur[2]}`
-        }
-
-        return `<div id=task-${task.id} class="tasks-section__task">
-                    <div class="handle">
-                        <i class="fas fa-ellipsis-v"></i>
-                    </div>
-                    <input type="checkbox">
-                    <div class="card-text">${taskStr}</div>
-                </div>`;
-    });
-
-    tasksContainer.innerHTML = tasksHtml.join('');
-};
-
-const updateTasksSection = async (listIds, queries = []) => {
-    let tasks = await getTasks(listIds);
-    for (let query of queries)
-        tasks = await filterTasks(tasks, query);
-    await displayTasks(tasks);
-}
-
-
-
-
-
 
 //---------------------------------------------------------------------
 
@@ -311,63 +221,20 @@ window.addEventListener('DOMContentLoaded', async (event) => {
     //grab the user's lists when DOM is loaded
     await fetchLists();
 
-    //------------------Rename Button Modal----------------------------------------------
-    // Get the modal
-    let modal_rename = document.querySelector(`.modal-rename`); //return a node list
-
-    // Get the button that opens the modal
-    let btn_rename = document.querySelectorAll(".list__btn--rename"); //return a node list
-
-    // Get the <span> element that closes the modal
-    let span_rename = document.getElementsByClassName("close-rename")[0];
-
-    // Get the <input> element that closes the modal
-    let btnCancel_rename = document.getElementsByClassName("list-cancel-rename")[0];
-
-    //define the rename btn unique id
-    let renameBtnId;
-
-    // When the user clicks on the button, open the modal
-    btn_rename.forEach(btn => {
-        btn.addEventListener('click', event => {
-            // console.log(event.target);
-            renameBtnId = event.target.classList[1];
-            event.stopPropagation();
-            modal_rename.style.display = "block";
-        })
-    })
-
-    // When the user clicks on <span> (x), close the modal
-    span_rename.addEventListener('click', event => {
-        event.stopPropagation();
-        modal_rename.style.display = "none";
-    })
-
-    // When the user clicks on <input> "cancel", close the modal
-    btnCancel_rename.addEventListener('click', event => {
-        event.stopPropagation();
-        modal_rename.style.display = "none";
-    })
-
-    // When the user clicks anywhere outside of the modal, close it
-    modal_rename.addEventListener('click', event => {
-        //event.target returns an html element
-        event.stopPropagation();
-        if (event.target === modal_rename) {
-            modal_rename.style.display = "none";
-        }
-    });
-
-    //---------------------------------------------------------------------------------------
-
     const listAnchors = document.querySelectorAll('.list-anchors');
     const inputAddListBtn = document.querySelector('.list-add');
     const inputListValue = document.querySelector('.list-name--input');
 
-    const inputRenameSingle = document.querySelector('.list-rename');
-    const inputRenameValue = document.querySelector('.list-name-rename--input');
+    //------------------Rename Button Modal----------------------------------------------
+    // Declare the modal
+    let modal_rename;
 
-    const deleteBtns = document.querySelectorAll('.list__btn--delete');
+    // Declare the <span> element that closes the modal
+    let span_rename;
+
+    // Declare the <input> element that closes the modal
+    let btnCancel_rename;
+    //---------------------------------------------------------------------------------
 
     //Listen for a click on each anchor element to display tasks
     listAnchors.forEach(list => {
@@ -375,7 +242,6 @@ window.addEventListener('DOMContentLoaded', async (event) => {
             event.preventDefault();
             event.stopPropagation();
             await updateTasksSection(event.target.id);
-            // await fetchOneList(event.target.id);
         })
     })
 
@@ -387,24 +253,60 @@ window.addEventListener('DOMContentLoaded', async (event) => {
         inputListValue.value = "";
     });
 
-    // listen for a click on the save btn to rename the list
-    /*There is a bug where the modal will not pop up anymore when the
-    rename btn is clicked only after creating a new list. */
-    inputRenameSingle.addEventListener('click', async (event) => {
-        event.stopPropagation();
-        event.preventDefault();
+    //----------------------------------------------------------------------------------
 
-        await updateList(renameBtnId, inputRenameValue.value);
-        modal_rename.style.display = "none";
-        inputRenameValue.value = "";
-    });
+    const rootDiv = document.getElementById("lists")
+    let modalId;
 
-    //Listen for a click on the delete btn
-    deleteBtns.forEach( btn => {
-        btn.addEventListener('click', async (event) => {
-            const deleteId = event.target.classList[1];
+    rootDiv.addEventListener('click', async (e) => {
+        e.preventDefault()
+        let btn = e.target
+        // console.log("event.target btn: ", btn.classList)
+
+        //Delete a list when "delete" btn is clicked
+        if (btn.classList.contains('list__btn--delete')) {
+            let deleteId = e.target.classList[1];
             await deleteList(deleteId);
-        })
-    });
+        }
 
+        //Display the modal when the "rename" btn is clicked
+        if (btn.classList.contains('list__btn--rename')) {
+            modalId = e.target.classList[1];
+
+            //we can reassign the buttons
+            modal_rename = document.querySelector(`#myModal-rename-${modalId}`);
+            span_rename = document.querySelector(`#close-rename-${modalId}`);
+            btnCancel_rename = document.querySelector(".list-cancel-rename");
+
+            modal_rename.style.display = 'block';
+        }
+
+        //Rename the targeted list when "save" btn is clicked
+        if (btn.classList.contains("list-rename")) {
+
+            let inputVal = document.querySelector(`#rename-input-${modalId}`);
+            // console.log("id:", modalId);
+            // console.log(inputVal.value);
+
+            await updateList(modalId, inputVal.value);
+
+            modal_rename.style.display = "none";
+            inputVal.value = "";
+        }
+
+        //Exit out of the modal when the "X" btn is clicked
+        if (btn.classList.contains("close-rename")) {
+            modal_rename.style.display = "none";
+        }
+
+        //Exit out of the modal when the "cancel" btn is clicked
+        if (btn.classList.contains("list-cancel-rename")) {
+            modal_rename.style.display = "none";
+        }
+
+        //Exit out of the modal
+        if (btn.classList.contains("modal-rename")) {
+            modal_rename.style.display = "none";
+        }
+    }, true)
 });
